Transition to 1 Activity Pt. 4
Added the new autonomous and teleop buttons.
Added framework for showing whether a button is on or off.
Changed strings to all caps to fit with TORC style.

Note: things are put automatically as gone that will never be gone during app use are gone to ease working on lower sections due to automatic refreshing of the preview on the designer.