Simplify Screens
Changed all counters and checkboxes to buttons.
With a single tap, the buttons increase in value or toggle on.
With a long tap, a vibration happens as feedback and value decreases or button toggles off.
Switched Rows and Columns buttons in teleop, to make it easier to understand, in my opinion.
Changed string in activity_autonomous.xml due to removing of 'autonomous' and 'teleop' from start of cryptobox string names.