Started Work on Saving
Changed OWN JEWEL and OTHER JEWEL to red and blue jewel, and am asking for the alliance color of teh robot, makes scouting easier and will convert back to own jewel and other jewel for saving.
There are many errors in MainActivity.java, but they are being worked on.
The saveToCSV function is taken from TORC 2137's code, and is being changed to work on our much simpler program.