Fixed app to get on newer phones
It turns out, that newer Android phones need to request "dangerous" permissions in the app, not upon install. I coded (read copied from the internet) a whopping 6 lines of code to check if there is the permission (older phones) then ask if it doesn't have it. Since I am installing the app, I will always hit accept.

In the root folder for the project, I created a text document with a list of who's phones require moving the file to a different folder. (2017 Scouting App/ 2017 Scouting App for the blue laptop)